refactor: Suspense 및 ErrorBoundary 로직 개선
